LayoutTests/inspector/debugger/content-providers-expected.txt

Tests ContentProvider implementations.


Running: testScriptContentProvider
Request foo.js source.

Running: testConcatenatedScriptsContentProvider
Request foo.js source.
Request bar.js source.
Request baz.js source.

  <script>function foo() { return 0; }</script>               <script>function bar() { return 0; }</script>

            <script>
function baz()
{
  return 0;
}
</script>

Running: testResourceContentProvider
Request resource content.
